% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gdp_data.R
\name{load_gdp_data}
\alias{load_gdp_data}
\title{Load the democracy and GDP dataset}
\usage{
load_gdp_data()
}
\value{
A tibble containing democracy and GDP data.
}
\description{
Loads the merged democracy and GDP dataset bundled with the package.
}
