% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_unemployment_data.R
\name{load_unemployment_data}
\alias{load_unemployment_data}
\title{Load the democracy and unemployment dataset}
\usage{
load_unemployment_data()
}
\value{
A tibble containing democracy and unemployment data.
}
\description{
Loads the merged democracy and unemployment dataset bundled with the package.
}
