% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_female_leaders.R
\name{plot_female_leaders}
\alias{plot_female_leaders}
\title{Plot Top Countries by Years of Female Leadership}
\usage{
plot_female_leaders(data, n = 10)
}
\arguments{
\item{data}{A data frame from \code{load_gdp_data()}.}

\item{n}{Number of top countries to display. Default is 10.}
}
\value{
A ggplot object.
}
\description{
Creates a bar plot showing the top countries with the most years of
female national leadership.
}
