% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the female leadership dataset}
\usage{
load_data()
}
\value{
A tibble containing the female leadership data.
}
\description{
Loads the raw dataset directly from the TidyTuesday GitHub repository.
}
\examples{
dat <- load_data()
}
